requirements.txt
cachetools==5.3.1
deprecation==2.1.0
grpcio==1.54.2
packaging==23.1
protobuf==4.23.3
pydantic==1.10.9
python-dateutil==2.8.2
PyYAML==6.0
six==1.16.0
tinkoff==0.1.1
tinkoff-investments==0.2.0b58
typing_extensions==4.6.3
